+++ test/unit/test_hooks/test_negative_edge_sampler_hook.py
@@ -8,7 +8,7 @@ from tgm.hooks import NegativeEdgeSamplerHook
 
 @pytest.fixture
 def data():
-    edge_index = torch.LongTensor([[2, 2], [2, 4], [1, 8]])
+    edge_index = torch.IntTensor([[2, 2], [2, 4], [1, 8]])
     edge_timestamps = torch.LongTensor([1, 5, 20])
     return DGData.from_raw(edge_timestamps, edge_index)
